Fixed: All error, warnings and bugs found on main pages.
