+++ src/Components/EditProjectForm.jsx
@@ -34,23 +34,27 @@ export default function EditProjectForm(props){
       }
     }
 
-    axios.patch(`${prefix}projects/${id}`, {project} ,{headers: {"Authorization": `Bearer ${user.jwt}`}})
-    .then(res => {
-      console.log("Project was successfully edited!");
-      // Redirect to the project\
-      history.push(`/project/${res.data.id}`)
-    })
-    .catch(err => {
-      console.log("Project was NOT successfully edited!")
-      setForm({"name":"", "description":""});
-    })
+    if(user.jwt){
+      axios.patch(`${prefix}projects/${id}`, {project} ,{headers: {"Authorization": `Bearer ${user.jwt}`}})
+      .then(res => {
+        console.log("Project was successfully edited!");
+        // Redirect to the project\
+        history.push(`/project/${res.data.id}`)
+      })
+      .catch(err => {
+        console.log("Project was NOT successfully edited!")
+        setForm({"name":"", "description":""});
+      })
+    }
   }
 
   // Delete project
   function deleteProject(e){
     e.preventDefault();
 
-    axios.delete(`${prefix}projects/${id}` ,{headers: {"Authorization": `Bearer ${user.jwt}`}})
+    if(user.jwt){
+      axios.delete(`${prefix}projects/${id}` ,{headers: {"Authorization": `Bearer ${user.jwt}`}})
+    }
 
   }
