+++ src/Components/ProjectUsersTable.jsx
@@ -44,10 +44,13 @@ export default function ProjectUsersTable(props){
             }
         ]
     }
-    axios.delete(`${prefix}projects/${id}/users`, {project}, {headers: {"Authorization": `Bearer ${user.jwt}`}})
-    .then(res => res.body)
-    .then(body => console.log(body))
-    .catch(err => console.log(err))
+
+    if(user.jwt){
+        axios.delete(`${prefix}projects/${id}/users`, {project}, {headers: {"Authorization": `Bearer ${user.jwt}`}})
+        .then(res => res.body)
+        .then(body => console.log(body))
+        .catch(err => console.log(err))
+    }
   }
 
   return (
