+++ src/Components/SampleUserModal.jsx
@@ -1,15 +1,38 @@
 import Modal from 'react-bootstrap/Modal'
+import { useContext } from 'react'
+import {UserContext} from "./UserProvider"
 import { Link } from 'react-router-dom'
 import Admin from "./../assets/admin.svg"
 import Developer from "./../assets/developer.svg"
 import Client from "./../assets/pleb.svg"
 import Owner from "./../assets/owner.svg"
+import axios from 'axios'
 
 export default function SampleUserModal(props) {
 
+  const { userLogin,  prefix } = useContext(UserContext)
+
+  //Handle sign in.
+  function logInSample(e){
+    e.preventDefault();
+
+    let user = {
+      email: "sample@sample.com",
+      password: "password"
+    }
+
+    axios.post(`${prefix}users/signin`, { user })
+    .then(res => {
+      userLogin(res.data.username, res.data.email, res.data.jwt);
+    })
+    .catch(err => {
+      console.log(err);
+    })
+  }
+
   return (
     <Modal
-      {...props}s
+      {...props}
       size="lg"
       centered
     >
@@ -21,19 +44,19 @@ export default function SampleUserModal(props) {
       <Modal.Body>
         <div className="user">
           <img src={Client} alt="Client User"/> 
-          <Link className="btn btn-primary w-75" onClick={props.handleLogin} to="/dashboard">Log in Client</Link>
+          <Link className="btn btn-primary w-75" onClick={logInSample} to="/dashboard">Log in Client</Link>
         </div>
         <div className="user">
           <img src={Developer} alt="Client User"/>
-          <Link className="btn btn-primary w-75" onClick={props.handleLogin} to="/dashboard">Log in Developer</Link>
+          <Link className="btn btn-primary w-75" onClick={logInSample} to="/dashboard">Log in Developer</Link>
         </div>
         <div className="user">
           <img src={Admin} alt="Client User"/>
-          <Link className="btn btn-primary w-75" onClick={props.handleLogin} to="/dashboard">Log in Admin</Link>
+          <Link className="btn btn-primary w-75" onClick={logInSample} to="/dashboard">Log in Admin</Link>
         </div>
         <div className="user">
           <img src={Owner} alt="Client User"/>
-          <Link className="btn btn-primary w-75" onClick={props.handleLogin} to="/dashboard">Log in Owner</Link>
+          <Link className="btn btn-primary w-75" onClick={logInSample} to="/dashboard">Log in Owner</Link>
         </div>
       </Modal.Body>
     </Modal>
